refactor: Update TestTopology package structure and simplify topology submission
